+++ src/components/Header/Header.module.css
@@ -3,8 +3,7 @@
   background-color: green;
   display: flex;
 }
-.header img {
-}
+
 .text {
   font-size: 20px;
   font-family: Arial, Helvetica, sans-serif;
