+++ src/index.js
@@ -1,8 +1,31 @@
+import React from "react";
+import ReactDOM from "react-dom/client";
+import { BrowserRouter } from "react-router-dom";
 import reportWebVitals from "./reportWebVitals";
 
-import { rerenderEntireTree } from "./render";
 import { state } from "./redux/state";
 
+import App from "./App";
+import addPost from "./redux/state";
+import updateNewPostText from "./redux/state";
+
+import "./index.css";
+
+const root = ReactDOM.createRoot(document.getElementById("root"));
+
+const rerenderEntireTree = (state) => {
+  root.render(
+    <React.StrictMode>
+      <BrowserRouter>
+        <App
+          state={state}
+          addPost={addPost}
+          updateNewPostText={updateNewPostText}
+        />
+      </BrowserRouter>
+    </React.StrictMode>
+  );
+};
 rerenderEntireTree(state);
 
 // If you want to start measuring performance in your app, pass a function
