--- src/render.js
@@ -1,25 +1,0 @@
-import React from "react";
-import ReactDOM from "react-dom/client";
-import { BrowserRouter } from "react-router-dom";
-
-import App from "./App";
-import addPost from "./redux/state";
-import updateNewPostText from "./redux/state";
-
-import "./index.css";
-
-const root = ReactDOM.createRoot(document.getElementById("root"));
-
-export const rerenderEntireTree = (state) => {
-  root.render(
-    <React.StrictMode>
-      <BrowserRouter>
-        <App
-          state={state}
-          addPost={addPost}
-          updateNewPostText={updateNewPostText}
-        />
-      </BrowserRouter>
-    </React.StrictMode>
-  );
-};
